packed_ref_store: new class, extracted from `files_ref_store`
Extract a new `ref_store` class, `packed_ref_store`, for dealing with
references stored in a `packed-refs` file. Change `files_ref_store` to
delegate to a `packed_ref_store` for its interactions with the
`packed-refs` file.

`packed_ref_store` implements the `ref_store` API, but it has the
following limitations:

* It cannot store symbolic refs
* It cannot store reflogs
* It does not support locking individual references, though the whole
  `ref_store` can be locked (even outside of a transaction). It
  implements this by creating a `packed-refs.lock` file.
* It does not support renaming refs (though it would be possible to
  add this functionality)

Note also that the references stored in a `packed_refs_store` might
point at objects that no longer exist in the ODB. (This is not a
problem provided there is a corresponding loose reference with the
same name that hides the packed version of the reference.)

`files_ref_store` now contains a `packed_ref_store` instance that it
delegates to when it needs to interact with the `packed-refs` file.
The `files_ref_store` class is also responsible for coordinating
updates of packed and loose references in such a way that the locking
protocol is adhered to, as well as repacking loose objects into the
`packed_ref_store` when requested.

Aside from better separation of concerns, `packed_ref_store`
(together with changes to `files_ref_store`) implements a few
improvements relative to the old monolithic `files_ref_store`:

* New content is not staged to `packed-refs.lock` but rather to
  another temporary file called `packed-refs.new`. This allows the
  `packed-refs.lock` file to be kept around for longer than a
  transaction, which enables the following long-standing race
  conditions to be fixed.

* The old reference transaction code used to commit all loose
  reference changes, and only then lock and rewrite the new
  `packed-refs` file. This meant that for a period of time after the
  deletion of a loose ref, the packed ref with the same name was
  visible to other processes. But the packed ref probably had a stale
  value, and could even have pointed at an object that has long since
  been garbage collected. The new code activates the new `packed-refs`
  content *after* committing loose reference creations and
  updates (important to reduce the time that objects might temporarily
  become unreachable), but *before* committing loose reference
  deletions (thereby fixing the race just described). Moreover, the
  sometimes costly work of writing the new `packed-refs` content to
  `packed-refs.new` is done *before* any loose reference changes have
  been committed, thereby minimizing the time it takes to commit all
  reference updates and shortening the time that the references are in
  between the old and new states.

* The old `pack-refs` code released the `packed-refs` lock immediately
  after writing the newly packed refs, but not while pruning the loose
  references. This theoretically allowed some (admittedly very
  improbably) races such as the following:

  1. `pack-refs` reads value `A` for reference `R` and writes that
     value to the new new `packed-refs` file.
  2. Another process deletes reference `R`, which includes deleting it
     from the `packed-refs` file.
  3. Yet another process re-creates reference `R` with the same value `A`.
  4. The original `pack-refs` finally gets around to pruning `R`. It
     verifies that the current value of the packed ref is still `A`,
     so it deletes it.

  The net result would be that reference `R` disappears permanently
  instead of being left at value `A`. The new code fixes this race by
  retaining the lock on the `packed-refs` file during the pruning
  phase of `pack-refs`, too.

Note: I plan some more changes to `packed_ref_store` in future patch
series, and this refactoring will also make those changes a lot
easier.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>