read_packed_refs(): process peeled line along with the reference
Instead of processing one line per iteration of the main loop, process
a reference line plus its corresponding peeled line (if present).

Note that this change significantly tightens up the parsing of the
`parse-ref` file. Previously, any lines that were not recognized were
completely ignored. That's a dangerous policy, because it could lead
to silent data loss if the file gets corrupted. Instead, die if the
`packed-refs` file is not in the expected format.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>