read_packed_refs(): ensure that references are sorted when read
It doesn't actually matter now, because the references are only
iterated over to fill the associated `ref_cache`, which itself ensures
the correct order. But we want to get rid of the `ref_cache`, so we
want to be able to iterate directly over the `packed-refs` buffer, and
then the iteration will need to be ordered correctly.

In fact, we already write the `packed-refs` file sorted, but it is
possible that other Git clients don't get it right. So let's not
assume that a `packed-refs` file is sorted unless it is explicitly
declared to be so via a `sorted` trait in its header line.

If it is *not* declared to be sorted, then scan quickly through the
file to check. If it is found to be out of order, then sort the
records into a new memory-only copy rather than using the mmapped file
contents. This checking and sorting is done quickly, without parsing
the full file contents.

If the `packed-refs` file header contains the `sorted` trait, then
assume that the file is already sorted properly, and us the mmapped
file contents directly. Since we write the file correctly sorted,
include that trait when we write or rewrite a `packed-refs` file.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>