packed_ref_store: get rid of the `ref_cache` entirely
Now that everything has been written to read what it needs directly
out of the `packed-refs` file, `packed_ref_store` doesn't need to
maintain a `ref_cache` at all, so get rid of it.

First of all, this will save a lot of memory. Instead of complicated
parsed data structures in memory, we just mmap the file (potentially
sharing memory with other processes) and parse only what we need.

Moreover, since the mmapped access to the file is smart enough to read
only the parts of the file that it needs, this might save reading all
of the data from disk at all (at least if the file starts out sorted).

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>